+++ lib/matchers/document/be_stored_in.rb
@@ -1,0 +1,22 @@
+module Mongoid
+  module Matchers
+    class BeStoredInMatcher
+      def initialize(collection_name)
+        @collection_name = collection_name.to_s
+      end
+
+      def matches?(klass)
+        @klass = klass.is_a?(Class) ? klass : klass.class
+        @klass.collection_name == @collection_name
+      end
+
+      def description
+        "be stored in #{@collection_name.inspect}"
+      end
+    end
+
+    def be_stored_in(collection_name)
+      BeStoredInMatcher.new(collection_name)
+    end
+  end
+end